Move common methods to BraviaTVEntity
