+++ homeassistant/components/braviatv/entity.py
@@ -1,4 +1,6 @@
 """A entity class for BraviaTV integration."""
+from typing import Any
+
 from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
@@ -9,6 +11,8 @@ from .const import ATTR_MANUFACTURER, DOMAIN
 class BraviaTVEntity(CoordinatorEntity[BraviaTVCoordinator]):
     """BraviaTV entity class."""
 
+    _attr_has_entity_name = True
+
     def __init__(
         self,
         coordinator: BraviaTVCoordinator,
@@ -18,9 +22,18 @@ class BraviaTVEntity(CoordinatorEntity[BraviaTVCoordinator]):
         """Initialize the entity."""
         super().__init__(coordinator)
 
+        self._attr_unique_id = unique_id
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, unique_id)},
             manufacturer=ATTR_MANUFACTURER,
             model=model,
             name=f"{ATTR_MANUFACTURER} {model}",
         )
+
+    async def async_turn_on(self, **kwargs: Any) -> None:
+        """Turn the device on."""
+        await self.coordinator.async_turn_on()
+
+    async def async_turn_off(self, **kwargs: Any) -> None:
+        """Turn the device off."""
+        await self.coordinator.async_turn_off()
